backend/requirements.txt
fastapi
uvicorn[standard]
httpx
celery[redis]
redis
sqlalchemy
psycopg2-binary
python-dotenv
openai
pyjwt
pydantic
alembic
python-multipart
python-jose
PyJWT[crypto]
cryptography
